feat: Add session editing functionality with new UI and validation
- Introduce editable session times, including start and optional end times, via `AlertDialog` in `SessionHistoryScreen`.
- Add date and time pickers for improved UX and validation of inputs.
- Ensure validation for invalid formats, future dates, and logical time sequencing.
- Update `SessionViewModel` and `SessionRepository` to support session updates.
- Localize new session editing strings in `strings.xml` and `strings-fr.xml`.

+++ app/src/main/java/fr/centuryspine/lsgscores/data/session/SessionDao.kt
@@ -19,7 +19,7 @@ interface SessionDao {
     suspend fun getAllList(): List<Session>
 
     @Query("SELECT * FROM sessions WHERE id = :id")
-    fun getById(id: Int): Flow<Session?>
+    fun getById(id: Long): Flow<Session?>
 
     @Insert
     suspend fun insert(session: Session): Long
